+++ src/DefaultLabel.js
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export default (props) => (
+  <label {...props}>{props.children }</label>
+);