+++ src/defaultOptionToChoice.js
@@ -1,0 +1,7 @@
+export default (item) => {
+  if (item === null) return null;
+  if (typeof item === 'object') {
+    return { ...item };
+  }
+  return item;
+};